+++ tests/filters/test_colorize.py
@@ -7,15 +7,16 @@
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
-
+import tornado
 from preggy import expect
 
 from tests.base import FilterTestCase
 
 
 class ColorizeFilterTestCase(FilterTestCase):
+    @tornado.testing.gen_test
     def test_colorize_filter(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.colorize', 'colorize(40,80,20,ffffff)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.colorize', 'colorize(40,80,20,ffffff)')
         expected = self.get_fixture('colorize.jpg')
 
         ssim = self.get_ssim(image, expected)
