+++ tests/filters/test_fill.py
@@ -7,21 +7,21 @@
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
-
-
+import tornado
 from preggy import expect
 
 from tests.base import FilterTestCase
 
 
 class FillFilterTestCase(FilterTestCase):
+    @tornado.testing.gen_test
     def test_fill_filter_with_fixed_color(self):
         def config_context(context):
             context.request.fit_in = True
             context.request.width = 800
             context.request.height = 800
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.fill',
             'fill(0000ff)',
@@ -33,13 +33,14 @@ class FillFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.97)
 
+    @tornado.testing.gen_test
     def test_fill_filter_with_average(self):
         def config_context(context):
             context.request.fit_in = True
             context.request.width = 800
             context.request.height = 800
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.fill',
             'fill(auto)',
@@ -51,6 +52,7 @@ class FillFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.97)
 
+    @tornado.testing.gen_test
     def test_fill_filter_with_full_fit_in(self):
         def config_context(context):
             context.request.fit_in = True
@@ -58,7 +60,7 @@ class FillFilterTestCase(FilterTestCase):
             context.request.width = 800
             context.request.height = 800
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.fill',
             'fill(auto)',
@@ -70,6 +72,7 @@ class FillFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.97)
 
+    @tornado.testing.gen_test
     def test_fill_filter_with_full_fit_in_orig_width(self):
         def config_context(context):
             context.request.fit_in = True
@@ -77,7 +80,7 @@ class FillFilterTestCase(FilterTestCase):
             context.request.width = 'orig'
             context.request.height = 800
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.fill',
             'fill(auto)',
