+++ tests/filters/test_focal.py
@@ -7,16 +7,16 @@
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
-
-
+import tornado
 from preggy import expect
 
 from tests.base import FilterTestCase
 
 
 class FocalFilterTestCase(FilterTestCase):
+    @tornado.testing.gen_test
     def test_focal_filter_valid(self):
-        self.get_filtered(
+        yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.focal',
             'focal(146x156:279x208)'
@@ -28,12 +28,13 @@ class FocalFilterTestCase(FilterTestCase):
         expect(self.context.request.focal_points[0].y).to_equal(182)
         expect(self.context.request.focal_points[0].x).to_equal(212)
 
+    @tornado.testing.gen_test
     def test_focal_filter_no_change(self):
         def config_context(context):
             context.request.width = 400
             context.request.height = 600
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.focal',
             'focal(0x0:400x600)'
@@ -43,8 +44,9 @@ class FocalFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.99)
 
+    @tornado.testing.gen_test
     def test_focal_filter_pass_through_empty(self):
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.focal',
             'focal()'
@@ -54,8 +56,9 @@ class FocalFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.99)
 
+    @tornado.testing.gen_test
     def test_focal_filter_pass_through_invalid(self):
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.focal',
             'focal(0x0:0x0)'
@@ -65,8 +68,9 @@ class FocalFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.99)
 
+    @tornado.testing.gen_test
     def test_focal_filter_pass_through_dumb(self):
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.focal',
             'focal(lkasjdklfjlkajsfd)'
