+++ tests/filters/test_format.py
@@ -7,17 +7,19 @@
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
-
+import tornado
 from preggy import expect
 
 from tests.base import FilterTestCase
 
 
 class FormatFilterTestCase(FilterTestCase):
+    @tornado.testing.gen_test
     def test_invalid_format_should_be_null(self):
-        self.get_filtered('source.jpg', 'thumbor.filters.format', 'format(invalid)')
+        yield self.get_filtered('source.jpg', 'thumbor.filters.format', 'format(invalid)')
         expect(self.context.request.format).to_be_null()
 
+    @tornado.testing.gen_test
     def test_can_set_proper_format(self):
-        self.get_filtered('source.jpg', 'thumbor.filters.format', 'format(webp)')
+        yield self.get_filtered('source.jpg', 'thumbor.filters.format', 'format(webp)')
         expect(self.context.request.format).to_equal('webp')
