+++ tests/filters/test_redeye.py
@@ -7,16 +7,16 @@
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
-
-
+import tornado
 from preggy import expect
 from tests.base import FilterTestCase
 from thumbor.point import FocalPoint
 
 
 class RedEyeFilterTestCase(FilterTestCase):
+    @tornado.testing.gen_test
     def test_redeye_passthrough_when_no_face_points(self):
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'redeye.png',
             'thumbor.filters.redeye',
             'red_eye()'
@@ -27,6 +27,7 @@ class RedEyeFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.99)
 
+    @tornado.testing.gen_test
     def test_redeye_applied(self):
         expected = self.get_fixture('redeye_applied.png')
 
@@ -36,7 +37,7 @@ class RedEyeFilterTestCase(FilterTestCase):
 
             context.request.focal_points = [point]
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'redeye.png',
             'thumbor.filters.redeye',
             'red_eye()',
