+++ tests/filters/test_upscale.py
@@ -7,21 +7,21 @@
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
-
-
+import tornado
 from preggy import expect
 
 from tests.base import FilterTestCase
 
 
 class UpscaleFilterTestCase(FilterTestCase):
+    @tornado.testing.gen_test
     def test_upscale_filter_with_fit_in_big(self):
         def config_context(context):
             context.request.fit_in = True
             context.request.width = 1000
             context.request.height = 1000
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.upscale',
             'upscale()',
@@ -33,13 +33,14 @@ class UpscaleFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.97)
 
+    @tornado.testing.gen_test
     def test_upscale_filter_with_fit_in_small(self):
         def config_context(context):
             context.request.fit_in = True
             context.request.width = 400
             context.request.height = 400
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.upscale',
             'upscale()',
@@ -51,6 +52,7 @@ class UpscaleFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.97)
 
+    @tornado.testing.gen_test
     def test_upscale_filter_with_full_fit_in(self):
         def config_context(context):
             context.request.fit_in = True
@@ -58,7 +60,7 @@ class UpscaleFilterTestCase(FilterTestCase):
             context.request.width = 800
             context.request.height = 800
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.upscale',
             'upscale()',
@@ -70,6 +72,7 @@ class UpscaleFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.97)
 
+    @tornado.testing.gen_test
     def test_upscale_filter_with_adaptive_fit_in_big(self):
         def config_context(context):
             context.request.fit_in = True
@@ -77,7 +80,7 @@ class UpscaleFilterTestCase(FilterTestCase):
             context.request.width = 1000
             context.request.height = 1200
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.upscale',
             'upscale()',
@@ -89,6 +92,7 @@ class UpscaleFilterTestCase(FilterTestCase):
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.97)
 
+    @tornado.testing.gen_test
     def test_upscale_filter_with_adaptive_full_fit_in_big(self):
         def config_context(context):
             context.request.fit_in = True
@@ -97,7 +101,7 @@ class UpscaleFilterTestCase(FilterTestCase):
             context.request.width = 800
             context.request.height = 800
 
-        image = self.get_filtered(
+        image = yield self.get_filtered(
             'source.jpg',
             'thumbor.filters.upscale',
             'upscale()',
