+++ tests/filters/test_watermark.py
@@ -7,7 +7,7 @@
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
-
+import tornado
 from preggy import expect
 
 from tests.base import FilterTestCase
@@ -16,69 +16,80 @@ from tests.fixtures.watermark_fixtures import SOURCE_IMAGE_SIZES, WATERMARK_IMAG
 
 
 class WatermarkFilterTestCase(FilterTestCase):
+    @tornado.testing.gen_test
     def test_watermark_filter_centered(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,center,center,60)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,center,center,60)')
         expected = self.get_fixture('watermarkCenter.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_centered_x(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,center,40,20)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,center,40,20)')
         expected = self.get_fixture('watermarkCenterX.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_centered_y(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,80,center,50)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,80,center,50)')
         expected = self.get_fixture('watermarkCenterY.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_repeated(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,repeat,repeat,70)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,repeat,repeat,70)')
         expected = self.get_fixture('watermarkRepeat.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_repeated_x(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,repeat,center,70)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,repeat,center,70)')
         expected = self.get_fixture('watermarkRepeatX.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_repeated_y(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,30,repeat,30)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,30,repeat,30)')
         expected = self.get_fixture('watermarkRepeatY.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_detect_extension_simple(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark,30,-50,60)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark,30,-50,60)')
         expected = self.get_fixture('watermarkSimple.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_simple(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,30,-50,60)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,30,-50,60)')
         expected = self.get_fixture('watermarkSimple.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_calculated(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,4p,-30p,60)')
-        expected = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,32,-160,60)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,4p,-30p,60)')
+        expected = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,32,-160,60)')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_calculated_center(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,4p,center,60)')
-        expected = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,32,center,60)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,4p,center,60)')
+        expected = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,32,center,60)')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_calculated_repeat(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,repeat,30p,60)')
-        expected = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,repeat,160,60)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,repeat,30p,60)')
+        expected = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,repeat,160,60)')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
@@ -94,26 +105,30 @@ class WatermarkFilterTestCase(FilterTestCase):
 
             expect(filter.detect_and_get_ratio_position(pos, length)).to_be_equal_with_additional_info(expected, **test)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_simple_big(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermarkBig.png,-10,-100,50)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermarkBig.png,-10,-100,50)')
         expected = self.get_fixture('watermarkSimpleBig.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_simple_50p_width(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,30,-50,20,50)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,30,-50,20,50)')
         expected = self.get_fixture('watermarkResize50pWidth.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_simple_70p_height(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,30,-50,20,none,70)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,30,-50,20,none,70)')
         expected = self.get_fixture('watermarkResize70pHeight.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
 
+    @tornado.testing.gen_test
     def test_watermark_filter_simple_60p_80p(self):
-        image = self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,-30,-200,20,60,80)')
+        image = yield self.get_filtered('source.jpg', 'thumbor.filters.watermark', 'watermark(watermark.png,-30,-200,20,60,80)')
         expected = self.get_fixture('watermarkResize60p80p.jpg')
         ssim = self.get_ssim(image, expected)
         expect(ssim).to_be_greater_than(0.98)
@@ -163,3 +178,9 @@ class WatermarkFilterTestCase(FilterTestCase):
 
                     test['topic_name'] = 'image ratio'
                     expect(watermark_image).to_almost_equal(ratio, 2, **test)
+
+    @tornado.testing.gen_test
+    def test_watermark_filter_should_throw_exception(self):
+        with expect.error_to_happen(Exception):
+            yield self.get_filtered('source.jpg', 'thumbor.filters.watermark',
+                                    'watermark(not-existing-image.png,center,center,60)')
