+++ tests/test_transformer.py
@@ -8,51 +8,19 @@
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
 
-import tempfile
-import shutil
-from os.path import abspath, join, dirname
-
+import tornado
 from preggy import expect
+from tornado.testing import AsyncTestCase
 
 from thumbor.transformer import Transformer
-from thumbor.config import Config
-from thumbor.importer import Importer
-from thumbor.context import Context, ServerParameters
-from tests.base import TestCase
 from tests.fixtures.transformer_test_data import (
     TESTITEMS, MockSyncDetector, FIT_IN_CROP_DATA, TestData,
     MockErrorSyncDetector,
 )
 
 
-class TransformerTestCase(TestCase):
-    @classmethod
-    def setUpClass(cls, *args, **kw):
-        cls.root_path = tempfile.mkdtemp()
-        cls.loader_path = abspath(join(dirname(__file__), '../fixtures/images/'))
-        cls.base_uri = "/image"
-
-    @classmethod
-    def tearDownClass(cls, *args, **kw):
-        shutil.rmtree(cls.root_path)
-
-    def setUp(self):
-        super(TransformerTestCase, self).setUp()
-        self.has_handled = False
-
-    def get_context(self):
-        cfg = Config(SECURITY_KEY='ACME-SEC')
-        cfg.LOADER = "thumbor.loaders.file_loader"
-        cfg.FILE_LOADER_ROOT_PATH = self.loader_path
-        cfg.STORAGE = "thumbor.storages.file_storage"
-        cfg.FILE_STORAGE_ROOT_PATH = self.root_path
-
-        importer = Importer(cfg)
-        importer.import_modules()
-        server = ServerParameters(8889, 'localhost', 'thumbor.conf', None, 'info', None)
-        server.security_key = 'ACME-SEC'
-        return Context(server, cfg, importer)
-
+class TransformerTestCase(AsyncTestCase):
+    @tornado.testing.gen_test
     def test_invalid_crop(self):
         data = TestData(
             source_width=800, source_height=600,
@@ -66,27 +34,30 @@ class TransformerTestCase(TestCase):
         engine = ctx.modules.engine
 
         trans = Transformer(ctx)
-        trans.transform(lambda: None)
+        yield trans.transform()
         expect(engine.calls['crop']).to_be_empty()
 
     def validate_resize(self, test_data):
         expect(test_data).to_be_resized()
         expect(test_data).to_be_cropped()
 
+    @tornado.testing.gen_test
     def test_can_resize_images(self):
         for item in TESTITEMS:
             context = item.to_context()
             trans = Transformer(context)
-            trans.transform(lambda: None)
+            yield trans.transform()
             self.validate_resize(item)
 
+    @tornado.testing.gen_test
     def test_can_resize_images_with_detectors(self):
         for item in TESTITEMS:
             context = item.to_context(detectors=[MockSyncDetector])
             trans = Transformer(context)
-            trans.transform(lambda: None)
+            yield trans.transform()
             self.validate_resize(item)
 
+    @tornado.testing.gen_test
     def test_can_resize_images_with_detection_error(self):
         test_data = TestData(
             source_width=800, source_height=600,
@@ -97,9 +68,10 @@ class TransformerTestCase(TestCase):
         )
         context = test_data.to_context(detectors=[MockErrorSyncDetector], ignore_detector_error=True)
         trans = Transformer(context)
-        trans.transform(lambda: None)
+        yield trans.transform()
         self.validate_resize(test_data)
 
+    @tornado.testing.gen_test
     def test_can_resize_images_with_detection_error_not_ignoring_it(self):
         test_data = TestData(
             source_width=800, source_height=600,
@@ -111,17 +83,19 @@ class TransformerTestCase(TestCase):
         context = test_data.to_context(detectors=[MockErrorSyncDetector], ignore_detector_error=False)
         trans = Transformer(context)
 
-        trans.transform(lambda: None)
+        with expect.error_to_happen(Exception, 'x'):
+            yield trans.transform()
 
         expect(test_data.engine.calls['resize']).to_length(0)
 
+    @tornado.testing.gen_test
     def test_can_fit_in(self):
         for (test_data, (width, height, should_resize)) in FIT_IN_CROP_DATA:
             context = test_data.to_context()
             engine = context.modules.engine
 
             trans = Transformer(context)
-            trans.transform(lambda: None)
+            yield trans.transform()
 
             expect(engine.calls['crop']).to_be_empty()
 
@@ -132,6 +106,7 @@ class TransformerTestCase(TestCase):
             else:
                 expect(engine.calls['resize']).to_be_empty()
 
+    @tornado.testing.gen_test
     def test_can_transform_meta_with_orientation(self):
         data = TestData(
             source_width=800, source_height=600,
@@ -147,9 +122,10 @@ class TransformerTestCase(TestCase):
         engine = ctx.modules.engine
 
         trans = Transformer(ctx)
-        trans.transform(lambda: None)
+        yield trans.transform()
         expect(engine.calls['reorientate']).to_equal(1)
 
+    @tornado.testing.gen_test
     def test_can_transform_with_flip(self):
         data = TestData(
             source_width=800, source_height=600,
@@ -163,10 +139,11 @@ class TransformerTestCase(TestCase):
         engine = ctx.modules.engine
 
         trans = Transformer(ctx)
-        trans.transform(lambda: None)
+        yield trans.transform()
         expect(engine.calls['horizontal_flip']).to_equal(1)
         expect(engine.calls['vertical_flip']).to_equal(1)
 
+    @tornado.testing.gen_test
     def test_can_extract_cover(self):
         data = TestData(
             source_width=800, source_height=600,
@@ -186,9 +163,10 @@ class TransformerTestCase(TestCase):
         engine = ctx.modules.engine
 
         trans = Transformer(ctx)
-        trans.transform(lambda: None)
+        yield trans.transform()
         expect(engine.calls['cover']).to_equal(1)
 
+    @tornado.testing.gen_test
     def test_get_target_dimensions(self):
         data = TestData(
             source_width=800, source_height=600,
@@ -202,5 +180,5 @@ class TransformerTestCase(TestCase):
         trans = Transformer(ctx)
         dimensions = trans.get_target_dimensions()
         expect(dimensions).to_equal((600, 400))
-        trans.transform(lambda: 1)
+        yield trans.transform()
         expect(dimensions).to_equal((600, 400))
