+++ thumbor/app.py
@@ -7,7 +7,6 @@
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
-import tornado.web
 import tornado.ioloop
 
 from thumbor.handlers.blacklist import BlacklistHandler
